Update address.tpl
All the text has been renamed to match the front office text in customer/user addresses. Now, let's proceed to make the elements in required fields usable in customer addresses as well.
Let's also unhide the VAT number. When it's required in the required fields, and if the customer user is not in a company, it will not show. Therefore, it can save their address since it's a required product.